+++ www/go/core/jmap/Entity.php
@@ -84,7 +84,9 @@ abstract class Entity  extends OrmEntity {
 			$this->change();
 		}
 
-		$this->checkFilesFolder();
+		if(static::supportsFiles()) {
+			$this->checkFilesFolder();
+		}
 
 		$this->saveTmpFiles();
 
@@ -147,6 +149,7 @@ abstract class Entity  extends OrmEntity {
 	 * @throws \GO\Base\Exception\AccessDenied
 	 */
 	private function checkFilesFolder($force = false) {
+
 		if(!self::$checkFilesFolder || empty($this->filesFolderId)) {
 			return true;
 		}
@@ -181,7 +184,7 @@ abstract class Entity  extends OrmEntity {
 	}
 
 	protected static function checkFiles() {
-		if(property_exists(static::class, 'filesFolderId') && Module::isInstalled('legacy', 'files')) {
+		if(static::supportsFiles()) {
 			$tables = static::getMapping()->getTables();
 			$table = array_values($tables)[0]->getName();
 
@@ -221,7 +224,7 @@ abstract class Entity  extends OrmEntity {
 	 * @return bool
 	 */
 	private static function supportsFiles() {
-		return property_exists(static::class, 'filesFolderId');
+		return property_exists(static::class, 'filesFolderId') && Module::isInstalled("legacy", "files");
 	}
 
 	/**