+++ www/go/core/views/extjs3/form/Dialog.js
@@ -68,7 +68,7 @@ go.form.Dialog = Ext.extend(go.Window, {
 			type: "submit",
 			handler: function() {
 				this.submit().catch((error) => {
-					console.error(error);
+					GO.errorDialog.show(error.message);
 				});
 			},
 			scope: this
@@ -110,7 +110,7 @@ go.form.Dialog = Ext.extend(go.Window, {
 					text: t("Save"),
 					handler: function() {
 						this.submit().catch(function(error) {
-							console.error(error);
+							GO.errorDialog.show(error.message);
 						});
 					},
 					scope: this
@@ -418,8 +418,9 @@ go.form.Dialog = Ext.extend(go.Window, {
 			return serverId;
 
 		}).catch(function(error) {
+
 			const firstError = me.showFirstInvalidField();
-			return Promise.reject(firstError);
+			return Promise.reject(firstError ? {message: firstError} : error);
 		}).finally(function() {
 			me.actionComplete();
 		})
@@ -442,7 +443,7 @@ go.form.Dialog = Ext.extend(go.Window, {
 			// 		console.warn(f);
 			// 	}
 			// });
-			return;
+			return false;
 		}