+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Postfixadmin: fixed cleanup script of deleted mailboxes
 - Email: Fixed red messages in sent items
 - Core: Fixed html editor styles in Compact theme
+- Email: Fixed broken base64 encoded message in rare case
 
 23-08-2021 6.5.73
 - Timeregistration: Do not clone invoice_id upon cloning a time entry