+++ CHANGELOG.md
@@ -11,11 +11,13 @@ Core: System Settings -> Modules redesigned and searchable.
 ----
 
 
+24-08-2021 6.5.74
 - Studio: Initial config had "id" for autoExpandColumn instead of "showID".
 - Core: Custom fields can be rendered as text in templates with entity.customFields.asText.foo
 - Postfixadmin: fixed cleanup script of deleted mailboxes
 - Email: Fixed red messages in sent items
 - Core: Fixed html editor styles in Compact theme
+- Email: Fixed broken base64 encoded message in rare case
 
 23-08-2021 6.5.73
 - Timeregistration: Do not clone invoice_id upon cloning a time entry